+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Introduce csv2db icon
+- Check for mandatory command line arguments (ER #14)
+- Exit program immediately if no files were found (ER #17)
+- More debug statements for file reading operations (ER #15)
+- Color coded output (can be turned off via $NO_COLOR) (ER #20)
+- Provide failing data record as debug output (ER #19)
+- More precise status output (ER #24)
+- Change default data type to VARCHAR(1000) (ER #26)
+- Return code to indicate data loading issue (ER #25)
+- Introduce SQL Server support (ER #11)
+
+### Fixed
+- Pass correct exit return codes (bug #18)
+- Autocommit is explicitly turned off (bug #22)
+- Record list cleared on error (bug #23)
+- Preserve correct order of column list (bug #21)
+
 ## [1.3.1] - 2019-07-28
 
 ### Fixed